Project/requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_spectacular==0.26.3
Pillow==10.0.0
Pillow==10.0.0
python-dotenv==1.0.0
